+++ test-packages/test-services-e2e/TripPin/microsoft-o-data-service-sample-trippin-in-memory-models-service/AirportsApi.ts
@@ -20,8 +20,6 @@ import {
   Time,
   EdmTypeField
 } from '@sap-cloud-sdk/odata-v4';
-import { BigNumber } from 'bignumber.js';
-import { Moment, Duration } from 'moment';
 export class AirportsApi<
   DeSerializersT extends DeSerializers = DefaultDeSerializers
 > implements EntityApi<Airports<DeSerializersT>, DeSerializersT>