+++ test-packages/test-services/v4/multiple-schemas-service/TestEntity4Api.ts
@@ -19,8 +19,6 @@ import {
   Time,
   EdmTypeField
 } from '@sap-cloud-sdk/odata-v4';
-import { BigNumber } from 'bignumber.js';
-import { Moment, Duration } from 'moment';
 export class TestEntity4Api<
   DeSerializersT extends DeSerializers = DefaultDeSerializers
 > implements EntityApi<TestEntity4<DeSerializersT>, DeSerializersT>