+++ test-packages/test-services/v4/test-service/TestEntityCircularLinkChildApi.ts
@@ -22,8 +22,6 @@ import {
   EdmTypeField,
   OneToOneLink
 } from '@sap-cloud-sdk/odata-v4';
-import { BigNumber } from 'bignumber.js';
-import { Moment, Duration } from 'moment';
 export class TestEntityCircularLinkChildApi<
   DeSerializersT extends DeSerializers = DefaultDeSerializers
 > implements